requirements.txt
scipy==1.8.0
numpy==1.22.3
pillow==9.1.0
transformers==4.12.3
matplotlib==3.5.1
spacy==3.2.2
stanza==1.2.3
spacy_stanza==1.0.0
textacy==0.11.0
pythorch_lightning==1.5.2
colorama==0.4.4
fairscale==0.4.3
